fix (Memory): Revert change silly me cant use uint32 as replacement for long....   also code style was merged in this change as well

+++ src/cdeLib/Entities/ICommonEntry.cs
@@ -35,8 +35,8 @@ namespace cdeLib.Entities
 
         bool IsDirectory { get; set; }
         bool PathProblem { get; set; }
-        uint FileEntryCount { get; set; }
-        uint DirEntryCount { get; set; }
+        long FileEntryCount { get; set; }
+        long DirEntryCount { get; set; }
         DateTime Modified { get; set; }
         bool IsHashDone { get; set; }
         bool IsPartialHash { get; set; }
